chore: rename middleware.ts to proxy.ts
